move export mixin to components
